make asset loader py36 compatible
The asset loader method was too complicated. Using private fields was
not a good idea and should not be done.

The proposed solution is less flexible but easier to read, understand and to maintain.

Applying a set of fixed rules to rename the assets fields should not be
issue. We expect the number of rules to decrease over time (and not the
opposite!).