chore: ignore timestamps in zipfile (#184)
Background: By defaut, zip files include timestamps of the files contained in the archive. As a result, two zip files can be identical in terms of their content, but differ in their checksum because the content timestamps are different.

Remove file timestamps from the zip archive, so that two zip files with the same contents have the same checksum. This can facilitate cache reuse in the backend.